requirements.txt
httpx>=0.27
pandas>=2.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
PySide6>=6.7
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic>=2.4
loguru>=0.7
tenacity>=8.3
bcrypt>=4.1
PyJWT>=2.9
alembic>=1.13
tqdm>=4.66
